requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==4.0
django-cors-headers==4.4.0
djangorestframework==3.15.1
django-rest-knox==5.0.1
idna==3.8
pillow==10.4.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0